New branch: 8 TeV

+++ bin/make_fake_lib.sh
@@ -2,7 +2,10 @@ cd HNCommonLeptonFakes/conf/
 make -f Makefile.StandAlone
 cd -
 cd $LQANALYZER_LIB_PATH
-rm libHNCommonLeptonFakes.so
+if [[ -f libHNCommonLeptonFakes.so ]];
+    then
+    rm libHNCommonLeptonFakes.so
+fi
 cp $LQANALYZER_DIR/HNCommonLeptonFakes/Root/libHNCommonLeptonFakes.so .
 cd $LQANALYZER_DIR
 make distclean
